EffSecSpawn - move to subsection of effects
