+++ src/main/java/ch/njol/skript/effects/sections/EffSecSpawn.java
@@ -1,4 +1,4 @@
-package ch.njol.skript.sections;
+package ch.njol.skript.effects.sections;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.bukkitutil.EntityUtils;
