+++ src/main/java/ch/njol/skript/expressions/ExprLastSpawnedEntity.java
@@ -14,7 +14,7 @@ import ch.njol.skript.lang.ExpressionType;
 import ch.njol.skript.lang.Literal;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.util.SimpleExpression;
-import ch.njol.skript.sections.EffSecSpawn;
+import ch.njol.skript.effects.sections.EffSecSpawn;
 import ch.njol.util.Kleenean;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
